New subclass of VDC -- ScholarDataverse.
Changes to support the scholar dataverse types. This is the addsite page. It is functional, but still a work in progress with respect to a few loose ends required before the next build.  (Support affiliation data for the dataverse too, add validation to the new fields.)

TBD is Edit Site Page and also some corresponding changes needed in the presentation layer.