Changes to support the Manage Classifications page. This is work in progress, so please don't mind the toString output on the page and partial functionality.
Note: this page contains the link in the upper right to add classifications. this link is working, so you can get to the Add Classifications page this way.

Also, you will need to manually add this to your browser address bar as I haven't entered it on the network options page.  

http://localhost:8080/dvn/networkAdmin/ManageClassificationsPage.xhtml.

Tah!