I commented out the styles that were being applied to the "Download as..." dropdown link menu on the StudyPage and replaced them new styles that removes the dynamic hover behavior and shows the link menu on the page in a static fashion in all browsers.
This is a fix for Ticket #68525 where the link menu was being displayed haphazardly up and down the StudyPage for files with 200 or more files.  This occurred in all browsers, so a new layout needed to be quickly implemented for this build.  Removing the styles that made it a dynamic dropdown was the quickest and easiest solution, that didn't require any redesign or major code changes to the markup.