Moved "no set available" error message on the EditHarvestSitePage, so that it would no longer break the form's table when rendered. (Ticket #68327)
Cleaned up the stylesheet and mark up on the StudyPage, in an attempt to fix the "Edit..." link from floating to the top of the page when a study is released. (Ticket #68094)

Indented sub-classifications in the select classification list on the AddSitePage and EditSitePage. Also added tree-view image next to classification name. (Ticket #71909)