Changed the color in the stylesheet to a darker grey, for the PromotionalLinkSearchBoxPage, to improve readability. (Ticket #68835)
Added icefaces theme styles to the datatables in the ManageCollectionsPage, for the table headers and rows, to make the UI more consistent with other admin-manage pages. (Ticket #68830)

Edited Bundle.properties to change a period to a colon on the ContactUsConfirmPage. (Ticket #68633)

The ContributorRequestSuccessPage had some major issues, which have been fixed. The styles weren't rendering because style class were being referenced in plain xhtml div's using "styleClass" instead of "class". Made text revisions based on Isabelle and Ann's suggestions. (Ticket #68166)

Add period to error message on the AddClassificationsPage and changed the message to be more consistent with other required fields.(Ticket #68051)

On the EditCollectionPage, added some example queries and provided common query fields in our DVN, next to the input field for dynamic collections, to give users an example of what is required. (Ticket #71912)